refactor(backend): move board-presence Redis KV helpers into board-presence-store.ts
Second half of the pubsub/index.ts split: the board-presence domain
carries ~350 lines of Redis KV bookkeeping (per-board seq counter,
durable FIFO history, proof-of-presence stamps + local-map fallback
and its cleanup timer, writer holder, session->board mapping) that has
nothing to do with the generic pub/sub channel mechanics extracted in
the previous commit. Move it into BoardPresenceStore
(pubsub/board-presence-store.ts); PubSub re-exposes every method
unchanged via thin delegation, including the `@internal` test hooks
(resetLocalBoardMembershipForTest, etc.) board-presence.test.ts relies
on. Pure code-location move, no behavior change.
